Keep: IsFilePath()整理
[patchunicode:#1114]


git-svn-id: https://svn.code.sf.net/p/sakura-editor/code/sakura/trunk2@4204 f7ce1907-e4c7-47ca-9f76-12c87ed2c91c

+++ sakura_core/util/file.h
@@ -27,7 +27,7 @@
 
 bool fexist(LPCTSTR pszPath); //!< ファイルまたはディレクトリが存在すればtrue
 
-bool IsFilePath( const wchar_t*, int*, int*, bool = true );
+bool IsFilePath( const wchar_t*, size_t*, size_t*, bool = true );
 bool IsFileExists(const TCHAR* path, bool bFileOnly = false);
 bool IsDirectory(LPCTSTR pszPath);	// 2009.08.20 ryoji
